requirements.txt
numpy==1.21.0
opencv-python==4.5.3.56
torch==1.9.0
matplotlib==3.4.2
Pillow==8.2.0
scipy==1.7.0
tqdm==4.61.1
scikit-learn==0.24.2
pathlib==1.0.1
os==0.1.4

